requirements.txt
Faker==2.0.0
PyMySQL==0.9.3
python-dateutil==2.8.0
PyYAML==5.1.2
six==1.12.0
SQLAlchemy==1.3.7
SQLAlchemy-Utils==0.34.1
text-unidecode==1.2
